feat: optimize code
